Migrate Discussions room action
