+++ apps/meteor/app/discussion/client/index.ts
@@ -1,4 +1,3 @@
 // Other UI extensions
 import './lib/messageTypes/discussionMessage';
 import './createDiscussionMessageAction';
-import './tabBar';
